# backend/docs/swagger.yaml
basePath: /
definitions:
...
  domain.SystemBaseConfig:
    properties:
      agentOfflineFailureCount:
        type: integer
      alertNotificationBatchSize:
        type: integer
      alertNotificationMaxRetryCount:
        type: integer
      alertNotificationRetryBaseSeconds:
        type: integer
      alertNotificationRetryMaxMinutes:
        type: integer
      alertNotificationTimeoutSeconds:
        type: integer
      appName:
        type: string
      appSubtitle:
        type: string
      created_at:
        type: string
      iconData:
        type: string
      loginName:
        type: string
      passwordResetCaptchaTtlMinutes:
        type: integer
      passwordResetCodeTtlMinutes:
        type: integer
      passwordResetRateLimitMinutes:
        type: integer
      passwordResetSendCooldownMinutes:
        type: number
      resourceAlertConsecutiveCount:
        type: integer
      resourceCriticalThreshold:
        type: integer
      resourceWarningThreshold:
        type: integer
      siteName:
        type: string
      updated_at:
        type: string
      wecomStateTtlMinutes:
        type: integer
    type: object
  domain.Task:
    properties:
      created_at:
        type: string
      createdBy:
        type: string
      errorMessage:
        type: string
      finished_at:
        type: string
      id:
        type: string
      payload:
        type: object
      status:
        type: string
      targetId:
        type: string
      targetType:
        type: string
      type:
        type: string
      updated_at:
        type: string
    type: object
...
  router.errorResponse:
    properties:
      error:
        type: string
      message:
        type: string
    type: object
...
  router.systemBaseConfigRequestDoc:
    properties:
      agentOfflineFailureCount:
        example: 3
        type: integer
      alertNotificationBatchSize:
        example: 50
        type: integer
      alertNotificationMaxRetryCount:
        example: 6
        type: integer
      alertNotificationRetryBaseSeconds:
        example: 30
        type: integer
      alertNotificationRetryMaxMinutes:
        example: 15
        type: integer
      alertNotificationTimeoutSeconds:
        example: 8
        type: integer
      appName:
        example: KVM Manager
        type: string
      appSubtitle:
        example: VIRTUALIZATION OPS
        type: string
      iconData:
        example: /favicon.svg
        type: string
      loginName:
        example: KVM Manager
        type: string
      passwordResetCaptchaTtlMinutes:
        example: 1
        type: integer
      passwordResetCodeTtlMinutes:
        example: 10
        type: integer
      passwordResetRateLimitMinutes:
        example: 5
        type: integer
      passwordResetSendCooldownMinutes:
        example: 0.5
        type: number
      resourceAlertConsecutiveCount:
        example: 3
        type: integer
      resourceCriticalThreshold:
        example: 85
        type: integer
      resourceWarningThreshold:
        example: 70
        type: integer
      siteName:
        example: KVM Manager
        type: string
    type: object
...
  router.vmConfigCDROMDoc:
    properties:
      bus:
        example: sata
        type: string
      connected:
        example: true
        type: boolean
      name:
        example: sda
        type: string
      path:
        example: /iso/ubuntu.iso
        type: string
    type: object
  router.vmConfigDiskDoc:
    properties:
      bus:
        example: virtio
        type: string
      bytes:
        example: 107374182400
        type: integer
      device:
        example: disk
        type: string
      name:
        example: vda
        type: string
      path:
        example: /var/lib/libvirt/images/prod-web-01.snap
        type: string
      pool:
        example: default
        type: string
      sourcePath:
        example: /var/lib/libvirt/images/prod-web-01.qcow2
        type: string
      type:
        example: file
        type: string
    type: object
  router.vmConfigDocResponse:
    properties:
      arch:
        example: x86_64
        type: string
      autostart:
        example: true
        type: boolean
      cdroms:
        items:
          $ref: '#/definitions/router.vmConfigCDROMDoc'
        type: array
      currentCpu:
        example: 4
        type: integer
      currentMemoryBytes:
        example: 8589934592
        type: integer
      description:
        example: 生产 Web 服务
        type: string
      disks:
        items:
          $ref: '#/definitions/router.vmConfigDiskDoc'
        type: array
      graphics:
        $ref: '#/definitions/router.vmConfigGraphicsDoc'
      hostCpu:
        example: 32
        type: integer
      hostMemoryBytes:
        example: 137438953472
        type: integer
      interfaces:
        items:
          $ref: '#/definitions/router.vmConfigInterfaceDoc'
        type: array
      maximumCpu:
        example: 8
        type: integer
      maximumMemoryBytes:
        example: 17179869184
        type: integer
      memoryStatsPeriod:
        example: 5
        type: integer
      name:
        example: prod-web-01
        type: string
      osType:
        example: linux
        type: string
      status:
        example: running
        type: string
      uuid:
        example: 550e8400-e29b-41d4-a716-446655440000
        type: string
      xml:
        example: <domain type='kvm'>...</domain>
        type: string
    type: object
  router.vmConfigGraphicsDoc:
    properties:
      listen:
        example: 0.0.0.0
        type: string
      passwordEnabled:
        example: true
        type: boolean
      port:
        example: "5901"
        type: string
      type:
        example: vnc
        type: string
    type: object
  router.vmConfigInterfaceDoc:
    properties:
      mac:
        example: "52:54:00:12:34:56"
        type: string
      model:
        example: virtio
        type: string
      name:
        example: vnet0
        type: string
      source:
        example: default
        type: string
      type:
        example: network
        type: string
    type: object
  router.vmConfigUpdateDocRequest:
    properties:
      currentCpu:
        example: 4
        type: integer
      currentMemoryMB:
        example: 8192
        type: integer
      description:
        example: 生产 Web 服务
        type: string
      maximumCpu:
        example: 8
        type: integer
      maximumMemoryMB:
        example: 16384
        type: integer
      memoryStatsPeriod:
        example: 5
        type: integer
    type: object
  router.vmConfigUpdateDocResponse:
    properties:
      config:
        $ref: '#/definitions/router.vmConfigDocResponse'
      task:
        $ref: '#/definitions/domain.Task'
    type: object
...
info:
  contact: {}
  description: KVM 虚拟化资源管理控制台后端 API，提供认证、Agent 管理、运行态资源、刷新事件、任务、审计和告警接口。
  title: KVM Manager API
  version: "1.0"
paths:
...
  /api/settings/base-config:
    get:
      description: 返回网站名称、登录页名称、控制台品牌名称、控制台品牌副标题、图标、安全时效、资源阈值、Agent 判定和告警通知策略配置。需要基础配置查看或管理权限。
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/domain.SystemBaseConfig'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/router.errorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/router.errorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/router.errorResponse'
      security:
      - BearerAuth: []
      summary: 获取基础配置
      tags:
      - settings
    put:
      consumes:
      - application/json
      description: 更新网站名称、登录/注册界面名称、控制台品牌名称、控制台品牌副标题、图标、安全时效、资源阈值、Agent 判定和告警通知策略配置。图标支持站内路径或图片
        Data URL，最大 512KB。
      parameters:
      - description: 基础配置
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/router.systemBaseConfigRequestDoc'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/domain.SystemBaseConfig'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/router.errorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/router.errorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/router.errorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/router.errorResponse'
      security:
      - BearerAuth: []
      summary: 更新基础配置
      tags:
      - settings
...
  /api/vms/{id}/config:
    get:
      description: 后端根据虚拟机所属 Agent 实时读取 libvirt 配置，Agent 使用当前 dumpxml 返回 CPU/内存上下限、介质、磁盘、网卡、自启动和描述。
      parameters:
      - description: 虚拟机 ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/router.vmConfigDocResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/router.errorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/router.errorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/router.errorResponse'
        "503":
          description: Service Unavailable
          schema:
            $ref: '#/definitions/router.errorResponse'
      security:
      - BearerAuth: []
      summary: 获取虚拟机真实配置
      tags:
      - realtime
    put:
      consumes:
      - application/json
      description: 修改虚拟机描述、vCPU 当前/最大分配和内存当前/最大分配。运行中的虚拟机支持在已预留上限内热扩容当前 CPU 与内存，最大
        CPU 和最大内存仍需关机后修改。后端转发到 Agent 执行 virsh setvcpus、setmaxmem、setmem 和 desc；运行中热扩容使用
        --live --config。
      parameters:
      - description: 虚拟机 ID
        in: path
        name: id
        required: true
        type: string
      - description: 虚拟机配置
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/router.vmConfigUpdateDocRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/router.vmConfigUpdateDocResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/router.errorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/router.errorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/router.errorResponse'
        "503":
          description: Service Unavailable
          schema:
            $ref: '#/definitions/router.errorResponse'
      security:
      - BearerAuth: []
      summary: 修改虚拟机资源配置
      tags:
      - realtime
...
securityDefinitions:
  BearerAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"